미해결: Image rendering 안됨
